Only the welcome page opens on startup
