+++ CorePreferences/src/au/gov/asd/tac/constellation/preferences/Bundle.properties
@@ -11,7 +11,7 @@ ApplicationOptionsPanel.autosaveCheckBox.text=Autosave every
 ApplicationOptionsPanel.autosaveLabel.text=minutes
 ApplicationOptionsPanel.startupPanel.AccessibleContext.accessibleName=Startup
 ApplicationOptionsPanel.startupPanel.border.title=Startup
-ApplicationOptionsPanel.startupTutorialCheckbox.text=Display tutorial panel on startup
+ApplicationOptionsPanel.startupTutorialCheckbox.text=Display welcome page on startup
 ApplicationOptionsPanel.displayPanel.AccessibleContext.accessibleName=Display
 ApplicationOptionsPanel.displayPanel.border.title=Display
 ApplicationOptionsPanel.freezeGraphCheckBox.text=Freeze Graph View
